struct/params: move Metric to pipe
